Tcrud v0.2 Response / NestedReply

+++ src/main/java/com/example/tcrud/service/BoardService.java
@@ -17,6 +17,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityNotFoundException;
 import java.security.Principal;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -62,9 +64,8 @@ public class BoardService {
         Page<Board> page = boardRepository.findAll(pageable);
         Page<BoardDto.BoardListDto> responseList =
                 page.map(board -> {
-                    BoardDto.BoardListDto boardListDto =
-                            modelMapper.map(board, BoardDto.BoardListDto.class);
-                    int replyCount = boardRepository.getReplyCountByBoardId(board.getId());
+                    BoardDto.BoardListDto boardListDto = modelMapper.map(board, BoardDto.BoardListDto.class);
+                    long replyCount = countBoardIdReplies(board.getId());
                     boardListDto.setReplycnt(replyCount);
                     return boardListDto;
                 });
@@ -101,7 +102,7 @@ public class BoardService {
         Page<BoardDto.BoardListDto> responseList =
                 page.map(board -> {
                     BoardDto.BoardListDto boardListDto = modelMapper.map(board, BoardDto.BoardListDto.class);
-                    int replyCount = boardRepository.getReplyCountByBoardId(board.getId());
+                    long replyCount = countBoardIdReplies(board.getId());
                     boardListDto.setReplycnt(replyCount);
                     return boardListDto;
                 });
@@ -114,7 +115,7 @@ public class BoardService {
         Page<BoardDto.BoardListDto> responseList =
                 page.map(board -> {
                     BoardDto.BoardListDto boardListDto = modelMapper.map(board, BoardDto.BoardListDto.class);
-                    int replyCount = boardRepository.getReplyCountByBoardId(board.getId());
+                    long replyCount = countBoardIdReplies(board.getId());
                     boardListDto.setReplycnt(replyCount);
                     return boardListDto;
                 });
@@ -128,7 +129,7 @@ public class BoardService {
         Page<BoardDto.BoardListDto> responseList =
                 page.map(board -> {
                     BoardDto.BoardListDto boardListDto = modelMapper.map(board, BoardDto.BoardListDto.class);
-                    int replyCount = boardRepository.getReplyCountByBoardId(board.getId());
+                    long replyCount = countBoardIdReplies(board.getId());
                     boardListDto.setReplycnt(replyCount);
                     return boardListDto;
                 });
@@ -141,7 +142,7 @@ public class BoardService {
         Page<BoardDto.BoardListDto> responseList =
                 page.map(board -> {
                     BoardDto.BoardListDto boardListDto = modelMapper.map(board, BoardDto.BoardListDto.class);
-                    int replyCount = boardRepository.getReplyCountByBoardId(board.getId());
+                    long replyCount = countBoardIdReplies(board.getId());
                     boardListDto.setReplycnt(replyCount);
                     return boardListDto;
                 });
@@ -154,7 +155,7 @@ public class BoardService {
         Page<BoardDto.BoardListDto> responseList =
                 page.map(board -> {
                     BoardDto.BoardListDto boardListDto = modelMapper.map(board, BoardDto.BoardListDto.class);
-                    int replyCount = boardRepository.getReplyCountByBoardId(board.getId());
+                    long replyCount = countBoardIdReplies(board.getId());
                     boardListDto.setReplycnt(replyCount);
                     return boardListDto;
                 });
@@ -185,33 +186,112 @@ public class BoardService {
         }
     }
 
-    //    TODO :---------------------- REPLY --------------------------------
+    //    TODO : 대댓글 LIST처리용 댓글카운트
+    public Long countBoardIdReplies(Long id)
+    {
+        return replyRepository.countByBoardId(id);
+    }
+
+//    TODO :---------------------- REPLY --------------------------------
+
+    //    REPLY LIST
+    public List<ReplyDto.Response> readBoardIdReplies(Long boardId) {
+//        List<Reply> rootReplies = replyRepository.findAllWithMemberAndParentByPostIdOrderByParentIdAscNullsFirstCommentIdAsc(boardId);
+        List<Reply> rootReplies = replyRepository.findAllByBoardId(boardId);
+        List<ReplyDto.Response> responseList = new ArrayList<>();
+
+        for (Reply rootReply : rootReplies) {
+            traverseReplyHierarchy(rootReply, responseList);
+        }
+
+        return responseList;
+    }
+
+
+    private void traverseReplyHierarchy(Reply reply, List<ReplyDto.Response> responseList) {
+        if (reply.getParent() == null) { // 최상위 댓글만 추가
+            ReplyDto.Response response = new ReplyDto.Response(reply);
+            if (!responseList.contains(response)) {
+                responseList.add(response);
+            }
+        }
+
+        List<Reply> children = reply.getChildren();
+        if (children != null) {
+            for (Reply child : children) {
+                traverseReplyHierarchy(child, responseList);
+            }
+        }
+    }
+
+
     // Reply CREATE
     public ReplyDto.Response createReply(ReplyDto.Request request) {
         Optional<Board> optionalBoard = boardRepository.findById(request.getBoardId());
         Optional<User> optionalUser = userRepository.findById(request.getUserId());
+
         if (!optionalBoard.isPresent() || !optionalUser.isPresent()) {
-            throw new EntityNotFoundException("정상적인 요청이 아닙니다.");
+            throw new EntityNotFoundException("게시글 또는 사용자를 찾을 수 없습니다.");
         }
+
+        Reply parentReply = null;
+
+        if (request.getParentReplyId() != null) {
+            Optional<Reply> optionalParentReply = replyRepository.findById(request.getParentReplyId());
+
+            if (!optionalParentReply.isPresent()) {
+                throw new EntityNotFoundException("댓글을 찾을 수 없습니다.");
+            }
+            parentReply = optionalParentReply.get();
+        }
+
         Board board = optionalBoard.get();
         User user = optionalUser.get();
-        Reply reply = request.toEntity(board, user);
+        Reply reply = request.toEntity(board, user, parentReply);
+
+
+        if (parentReply != null) {
+            reply.setParent(parentReply);
+            reply.setDepth(parentReply.getDepth() + 1);
+        } else {
+            reply.setDepth(0);
+        }
+
         replyRepository.save(reply);
         ReplyDto.Response response = new ReplyDto.Response(reply);
         return response;
     }
 
     //    Reply DELETE
-    public boolean removeReply(Long boardid, Long replyid) {
+    @Transactional
+    public boolean removeReply(Long boardid, Long replyid)
+    {
+        if ( crossCheckBR(boardid, replyid) )
+        {
+//            boardId가 존재하는지, replyId가 존재하는지, reply의 boardId가 매개변수와 일치하는지.
+//            board의 replyList에 replyId에 해당하는 댓글이 존재하는지
 
-        if (crossCheckBR(boardid, replyid)) {
-            replyRepository.deleteById(replyid);
+            Optional<Reply> getReply = replyRepository.findById(replyid);
+            Reply tempReply = getReply.get();
+
+            if(tempReply.getChildren() != null && !tempReply.getChildren().isEmpty())
+            {
+                tempReply.setContent("삭제된 댓글입니다");
+                tempReply.setHidden("Y");
+                replyRepository.save(tempReply);
+            }
+            else
+            {
+                replyRepository.deleteById(replyid);
+            }
             return true;
+
         } else {
             return false;
         }
     }
 
+    //    Reply UPDATE
     public ReplyDto.Response updateReply(Long id, ReplyDto.Request request) {
         Optional<Reply> optionalReply = replyRepository.findById(id);
         if (optionalReply.isPresent()) {
@@ -221,7 +301,7 @@ public class BoardService {
             ReplyDto.Response response = modelMapper.map(reply, ReplyDto.Response.class);
             return response;
         } else {
-            throw new RuntimeException("존재하지 않는 댓글입니다.");
+            throw new RuntimeException("reply is not exits!");
         }
     }
 
@@ -247,11 +327,7 @@ public class BoardService {
         }
     }
 
-    //    REPLY LIST
-    public Page<ReplyDto.Response> readBoardIdReplies(Long boardId, Pageable pageable) {
-        Page<Reply> replies = replyRepository.findByBoardId(boardId, pageable);
-        return replies.map(ReplyDto.Response::new);
-    }
+//    ----------------------------------------------------------------------------------
 
     //   TODO : 수정/삭제를 위한 로그인 사용자와 게시글 작성자가 일치하는지 확인. - BOARD
     public boolean isAuthorBoard(Long id, Principal principal) {
